refactored home.jsx
