+++ pages/home/index.js
@@ -1,0 +1,2 @@
+//TODO: clean up indirection
+module.exports = require('./HomePage.jsx');
